feat: rank test 코드 작성
