+++ src/test/java/lotto/LottoTest.java
@@ -1,13 +1,13 @@
 package lotto;
 
 
-import lotto.domain.*;
+import lotto.domain.LottoNumber;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
-import java.util.*;
+import java.util.HashSet;
+import java.util.Set;
 
-import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 public class LottoTest {
@@ -21,32 +21,4 @@ public class LottoTest {
         }).isInstanceOf(IllegalArgumentException.class);
     }
 
-    @DisplayName("일등 확인 테스트")
-    @Test
-    public void 일등확인() {
-        Lotto lotto = new Lotto(Arrays.asList(1, 2, 3, 4, 5, 6));
-        List<Lotto> lottoList = new ArrayList<>();
-        lottoList.add(lotto);
-
-        Lottos lottos = new Lottos(lottoList);
-        List<String> lastWeekWinner = Arrays.asList("1", "2", "3", "4", "5", "6");
-
-        ResultWinner winner = new ResultWinner();
-        assertThat(winner.countOfWinner(lottos, lastWeekWinner, "7").get(Rank.FIRST)).isEqualTo(1);
-    }
-
-    @DisplayName("이등 확인 테스트")
-    @Test
-    public void 이등확인() {
-        Lotto lotto = new Lotto(Arrays.asList(1, 2, 3, 4, 5, 6));
-
-        List<Lotto> lottoList = new ArrayList<>();
-        lottoList.add(lotto);
-
-        Lottos lottos = new Lottos(lottoList);
-        List<String> lastWeekWinner = Arrays.asList("1", "2", "3", "4", "5", "7");
-
-        ResultWinner winner = new ResultWinner();
-        assertThat(winner.countOfWinner(lottos, lastWeekWinner, "6").get(Rank.SECOND)).isEqualTo(1);
-    }
 }
